requirements.txt
# ton-client-py>=1.35.0.0
ton

aiogram>=2.21
python-dotenv>=0.20.0